Updates signup pane.
