+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/common/ModalBottomSheetContent.kt
@@ -2,28 +2,22 @@
 
 package com.stripe.android.financialconnections.features.common
 
-import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyListScope
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Divider
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
@@ -32,7 +26,6 @@ import com.stripe.android.financialconnections.R
 import com.stripe.android.financialconnections.model.DataAccessNotice
 import com.stripe.android.financialconnections.model.Image
 import com.stripe.android.financialconnections.model.LegalDetailsNotice
-import com.stripe.android.financialconnections.ui.ImageResource
 import com.stripe.android.financialconnections.ui.LocalImageLoader
 import com.stripe.android.financialconnections.ui.TextResource
 import com.stripe.android.financialconnections.ui.components.AnnotatedText
@@ -209,112 +202,36 @@ private fun ModalBottomSheetFooter(
     disclaimer: TextResource?,
     onConfirmModalClick: () -> Unit,
     cta: String
-) =
-    Column {
-        connectedAccountNotice?.let {
-            Spacer(modifier = Modifier.size(16.dp))
-            AnnotatedText(
-                modifier = Modifier.align(Alignment.CenterHorizontally),
-                text = it,
-                onClickableTextClick = onClickableTextClick,
-                defaultStyle = v3Typography.labelSmall.copy(
-                    color = v3Colors.textDefault,
-                    textAlign = TextAlign.Center
-                ),
-            )
-        }
-        disclaimer?.let {
-            Spacer(modifier = Modifier.size(16.dp))
-            AnnotatedText(
-                modifier = Modifier.align(Alignment.CenterHorizontally),
-                text = it,
-                onClickableTextClick = onClickableTextClick,
-                defaultStyle = v3Typography.labelSmall.copy(
-                    color = v3Colors.textDefault,
-                    textAlign = TextAlign.Center
-                ),
-            )
-        }
+) = Column {
+    connectedAccountNotice?.let {
         Spacer(modifier = Modifier.size(16.dp))
-        FinancialConnectionsButton(
-            onClick = { onConfirmModalClick() },
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Text(text = cta)
-        }
+        AnnotatedText(
+            modifier = Modifier.align(Alignment.CenterHorizontally),
+            text = it,
+            onClickableTextClick = onClickableTextClick,
+            defaultStyle = v3Typography.labelSmall.copy(
+                color = v3Colors.textDefault,
+                textAlign = TextAlign.Center
+            ),
+        )
     }
-
-@Composable
-internal fun ListItem(
-    bullet: BulletUI,
-    onClickableTextClick: (String) -> Unit,
-) {
-    val firstText = bullet.title ?: bullet.content ?: TextResource.Text("")
-    val secondText = remember(firstText) { bullet.content?.takeIf { bullet.title != null } }
-    val titleStyle = if (secondText != null) v3Typography.bodyMediumEmphasized else v3Typography.bodyMedium
-    Row {
-        ListItemIcon(icon = bullet.imageResource)
+    disclaimer?.let {
         Spacer(modifier = Modifier.size(16.dp))
-        Column {
-            AnnotatedText(
-                text = firstText,
-                defaultStyle = titleStyle.copy(color = v3Colors.textDefault),
-                onClickableTextClick = onClickableTextClick
-            )
-            secondText?.let {
-                AnnotatedText(
-                    text = requireNotNull(bullet.content),
-                    defaultStyle = v3Typography.bodySmall.copy(color = v3Colors.textSubdued),
-                    onClickableTextClick = onClickableTextClick
-                )
-            }
-        }
+        AnnotatedText(
+            modifier = Modifier.align(Alignment.CenterHorizontally),
+            text = it,
+            onClickableTextClick = onClickableTextClick,
+            defaultStyle = v3Typography.labelSmall.copy(
+                color = v3Colors.textDefault,
+                textAlign = TextAlign.Center
+            ),
+        )
     }
-}
-
-@Composable
-private fun ListItemIcon(icon: ImageResource?) {
-    val bulletColor = v3Colors.iconDefault
-    val iconSize = 20.dp
-    val modifier = Modifier
-        .size(iconSize)
-        .offset(y = 1.dp)
-    when (icon) {
-        // Render a bullet if no icon is provided
-        null -> Canvas(
-            modifier = modifier.padding((iconSize - 8.dp) / 2),
-            onDraw = { drawCircle(color = bulletColor) }
-        )
-
-        // Render the icon if it's a local resource
-        is ImageResource.Local -> Image(
-            modifier = modifier,
-            painter = painterResource(id = icon.resId),
-            contentDescription = null,
-        )
-
-        // Render the icon if it's a network resource, or fallback to a bullet if it fails to load
-        is ImageResource.Network -> StripeImage(
-            url = icon.url,
-            debugPainter = painterResource(id = R.drawable.stripe_ic_check_circle),
-            errorContent = {
-                Canvas(
-                    modifier = modifier.padding((iconSize - 8.dp) / 2),
-                    onDraw = { drawCircle(color = bulletColor) }
-                )
-            },
-            loadingContent = {
-                LoadingShimmerEffect { shimmer ->
-                    Spacer(
-                        modifier = modifier
-                            .clip(RoundedCornerShape(4.dp))
-                            .background(shimmer)
-                    )
-                }
-            },
-            imageLoader = LocalImageLoader.current,
-            contentDescription = null,
-            modifier = modifier
-        )
+    Spacer(modifier = Modifier.size(16.dp))
+    FinancialConnectionsButton(
+        onClick = { onConfirmModalClick() },
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        Text(text = cta)
     }
 }
